requirements.txt
alembic==1.2.1
boto3==1.9.241
botocore==1.12.241
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
idna==2.8
Inject==3.5.4
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
schematics==2.1.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.8
urllib3==1.25.6
Werkzeug==0.16.0
